% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_codes.R
\name{all_entities}
\alias{all_entities}
\title{Title}
\usage{
all_entities(spacy_out)
}
\arguments{
\item{spacy_out}{a spacy_out object}
}
\value{
list of named entities in texts
}
\description{
this function may be hidden in future
}

