% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-extractor-functions.R
\name{all_named_entities}
\alias{all_named_entities}
\title{Obtain a list of all named entities in spacy out data.table}
\usage{
all_named_entities(dt)
}
\arguments{
\item{dt}{a data table object from \code{"spacy_parse"} with 
named_entity option TRUE.}
}
\value{
a data.table of all named entities
}
\description{
Generate a list of all named entities from the named entity recoginition
in spaCy.
}
\examples{
\donttest{
txt <- "And now for something completely different, United States of America, President Barack Obama."
dt <- spacy_parse(txt, dependency = TRUE, named_entity = TRUE)
named_entities <- spacyr:::all_named_entities(dt)
dim(named_entities)
head(named_entities)
}
}

