% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{charvec_to_R}
\alias{charvec_to_R}
\title{Copy list of strings from Python to R character vector}
\usage{
charvec_to_R(name)
}
\arguments{
\item{name}{Python variable name}
}
\description{
Copy list of strings from Python to R character vector
}
\examples{

pyrun("l = ['a', 'b']")
pyrun("print(l)")
charvec_to_R("l")
pyrun("l2 = [u'a', u'b']")
charvec_to_R("l2")
}

