% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spacy.R
\name{check_spacy}
\alias{check_spacy}
\title{Checking the availability of spacy}
\usage{
check_spacy(which_python = NA)
}
\arguments{
\item{which_python}{specific python path you want to test the availability of
spacy. If you do not specify, this fuction looks for all python intallation
on the computer}
}
\description{
Checking the availablity of spaCy in the python on the system, as well as the
availability of spaCy model for English. This function has to be executed
before running \code{tag()}.
}
\section{Setting the path to the python executable}{
 Note that on some 
  systems, notably OS X / macOS, you may have installed a different version 
  of Python from that included in the base system.  OS X / macOS installs a 
  slightly older version of 2.7.x by default, for instance, in 
  \code{/usr/bin/python}.  Using homebrew, you may have installed a different
  version that gets placed in \code{/usr/local/bin}.  Even when this is 
  working at a command line (e.g. bash in the Terminal), when called from R 
  it may still look for \code{usr/bin/python}. This function finds the python
  executable with spaCy and set the path to it.
}
\author{
Akitaka Matsuo
}

