% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spacy.R
\name{check_spacy_c}
\alias{check_spacy_c}
\title{Checking the availability of spacy via Rcpp extension}
\usage{
check_spacy_c()
}
\description{
Checking the availablity of spaCy in the python on the system, as well as the
availability of spaCy model for English. This function has to be executed
before running \code{tag()}.
}
\author{
Tom Paskhalis
}

