% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{convert_to_data_table}
\alias{convert_to_data_table}
\title{Title}
\usage{
convert_to_data_table(tokens)
}
\arguments{
\item{tokens}{}
}
\value{
a data table of all variables in the data
}
\description{
Title
}

