% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_char_paragraph}
\alias{data_char_paragraph}
\title{A short paragraph of text for testing}
\format{An object of class \code{character} of length 1.}
\usage{
data_char_paragraph
}
\description{
A sample of text from the Irish budget debate of 2010 (531 tokens long).
}
\keyword{datasets}
