% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_char_sentences}
\alias{data_char_sentences}
\title{Sample short documents for testing}
\format{An object of class \code{character} of length 30.}
\usage{
data_char_sentences
}
\description{
A character object consisting of 30 short documents in plain text format for
testing.  Each document is one or two brief sentences.
}
\keyword{datasets}
