% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sentences}
\alias{data_sentences}
\title{sample short documents for testing}
\description{
Some sample short documents in plain text format for testing.  Each document is one or two short sentences.
}

