% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{docnames.tokenizedTexts_tagged}
\alias{docnames.tokenizedTexts_tagged}
\title{get or set document names}
\usage{
\method{docnames}{tokenizedTexts_tagged}(x)
}
\arguments{
\item{x}{the object with docnames}
}
\description{
Get or set the document names from a corpus or a document-feature matrix.
of the \link{dfm} object.
}
\keyword{internal}

