% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_consolidate.R
\name{entity_consolidate}
\alias{entity_consolidate}
\title{Consolidate named entities in parsed output}
\usage{
entity_consolidate(spacy_result, type = c("named", "extended", "all"))
}
\arguments{
\item{spacy_result}{a \code{data.frame} from \code{\link{spacy_parse}}.}

\item{type}{type of named entities to be consolidated, either \code{named}, \code{extended}, or
\code{all}.  See 
\url{https://spacy.io/docs/usage/entity-recognition#entity-types} for
details.}
}
\value{
A modified \code{data.table} of spacy outputs.
}
\description{
\code{entity_consolidate} concatinate named entity in a table of
results of \code{spacy_parse}
}
\examples{
\donttest{
spacy_initialize()

parsed_sentences <- spacy_parse(data_char_sentences, entity = TRUE, dependency = TRUE)
parsed_sentences_consolidated <- entity_consolidate(parsed_sentences)
head(parsed_sentences_consolidated, 30)
}
}
