% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-functions.R
\name{entity_extract}
\alias{entity_extract}
\title{Extract all named entities from parsed documents}
\usage{
entity_extract(spacy_result, type = c("named", "extended", "all"))
}
\arguments{
\item{spacy_result}{a \code{data.frame} from \code{\link{spacy_parse}}.}

\item{type}{type of named entities, either \code{named}, \code{extended}, or
\code{all}.  See 
\url{https://spacy.io/docs/usage/entity-recognition#entity-types} for
details.}
}
\value{
A \code{data.table} of all named entities, containing the following
  fields: \itemize{ \item{docname}{name of the documument a named entity is
  found} \item{entity}{the named entity} \item{entity_type}{type of named
  entities (e.g. PERSON, ORG, PERCENT, etc.)} }
}
\description{
\code{entity_extract} construct a table of all named entities from the
results of \code{spacy_parse}
}
\examples{
\donttest{
spacy_initialize()

parsed_sentences <- spacy_parse(data_char_sentences, entity = TRUE)
named_entities <- entity_extract(parsed_sentences)
head(named_entities, 30)
}
}
