% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_initialize.R
\name{find_spacy}
\alias{find_spacy}
\title{Find spaCy}
\usage{
find_spacy(model = "en")
}
\arguments{
\item{model}{name of the language model}
}
\value{
spacy_python
}
\description{
Locate the user's version of Python for which spaCy installed.
}
\keyword{internal}
