% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_dependency_data}
\alias{get_dependency_data}
\title{Title}
\usage{
get_dependency_data(spacy_out)
}
\arguments{
\item{spacy_out}{a spacy_out object}
}
\value{
data.frame of dependency relations
}
\description{
Title
}

