% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-extractor-functions.R
\name{get_ntokens_by_sent}
\alias{get_ntokens_by_sent}
\title{get the number of tokens by sentence from spacyr output}
\usage{
get_ntokens_by_sent(spacy_out)
}
\arguments{
\item{spacy_out}{a spacy_out object}
}
\value{
data.frame of dependency relations
}
\description{
get the number of tokens by sentence from spacyr output
}
\keyword{internal}
