% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-extractor-functions.R
\name{get_tags}
\alias{get_tags}
\title{tag parts of speech using spaCy via rPython}
\usage{
get_tags(spacy_out, tagset = c("google", "detailed"))
}
\arguments{
\item{spacy_out}{a spacy_out object}

\item{tagset}{character label for the tagset to use, either \code{"google"} 
or \code{"detailed"} to use the simplified Google tagset, or the more detailed 
scheme from the Penn Treebank (or the German Text Archive in case of German language model).}
}
\value{
a tokenized text object with tags
}
\description{
Tokenize a text using spaCy and tag the tokens with part-of-speech tags. 
Options exist for using either the Google or Detaled tagsets. See 
\url{http://spacy.io}.
}
\examples{
\donttest{
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
results <- spacy_parse(txt)
tokens <- tokens(results)
tokens_with_tag <- tokens_tag(tokens)
}
}
\keyword{internal}
