% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_tokens}
\alias{get_tokens}
\title{get tokens from spaCy output}
\usage{
get_tokens(spacy_out)
}
\arguments{
\item{spacy_out}{a spacy_out object}
}
\value{
a list of tokens
}
\description{
get tokens from spaCy output
}
\examples{
\donttest{
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
results <- parse_spacy(txt)
get_tags(results)
}
}

