% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spacy.R
\name{initialize_spacy_rpython}
\alias{initialize_spacy_rpython}
\title{Initialize spaCy via rPython}
\usage{
initialize_spacy_rpython(homebrew = FALSE)
}
\arguments{
\item{homebrew}{boolian for homebrewed version of spacy.}
}
\description{
Initialize spaCy via rPython
}
\author{
Akitaka Matsuo
}

