% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{ndoc.tokenizedTexts_tagged}
\alias{ndoc.tokenizedTexts_tagged}
\title{get the number of documents}
\usage{
\method{ndoc}{tokenizedTexts_tagged}(x)
}
\arguments{
\item{x}{a tokenizedTexts object}
}
\description{
\code{ndoc} returns an integer count of the number of documents
}
\keyword{internal}

