% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nounphrase-functions.R
\name{nounphrase_extract}
\alias{nounphrase_extract}
\alias{nounphrase_consolidate}
\title{Extract or consolidate entities from parsed documents}
\usage{
nounphrase_extract(x)

nounphrase_consolidate(x, concatenator = "_")
}
\arguments{
\item{x}{output from \code{\link{spacy_parse}}.}

\item{concatenator}{the character used to concatenator elements of multi-word
nounphrases.}
}
\value{
\code{noun} returns a \code{data.frame} of all named
  entities, containing the following fields: 
  \itemize{
  \item{\code{doc_id}}{ name of the documument containing the nounphrase} 
  \item{\code{sentence_id}}{ the sentence ID containing the nounphrase, within the document}
  \item{\code{nounphrase}}{ the nounphrase}
  \item{\code{root}}{ the root token of the nounphrase}
  }

\code{nounphrase_consolidate} returns a modified \code{data.frame} of
  parsed results, where the nounphrases have been combined into a single
  "token".  Currently, dependency parsing is removed when this consolidation
  occurs.
}
\description{
From an object parsed by \code{\link{spacy_parse}}, extract the nounphrase as a
separate object, or convert the multi-word nounphrases into single "token"
consisting of the concatenated elements of the multi-word nounphrases.
}
\examples{
\donttest{
spacy_initialize()

# entity extraction
txt <- "Mr. Smith of moved to San Francisco in December."
parsed <- spacy_parse(txt, nounphrase = TRUE)
entity_extract(parsed)
}
\donttest{
# consolidating multi-word nounphrases
txt <- "The House of Representatives voted to suspend aid to South Dakota."
parsed <- spacy_parse(txt, nounphrase = TRUE)
nounphrase_consolidate(parsed)
}
}
