% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{ntoken.tokenizedTexts_tagged}
\alias{ntoken.tokenizedTexts_tagged}
\alias{ntype.tokenizedTexts_tagged}
\title{count the number of tokens or types}
\usage{
\method{ntoken}{tokenizedTexts_tagged}(x, ...)

\method{ntype}{tokenizedTexts_tagged}(x, ...)
}
\arguments{
\item{x}{object whose tokens or types will be counted}

\item{...}{not used}
}
\description{
Return the count of tokens (total features) or types (unique features) in a
tokenizedText object
}
\keyword{internal}

