% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{numvec_to_R}
\alias{numvec_to_R}
\title{Get numeric vector from python __main__ namespace.}
\usage{
numvec_to_R(name)
}
\arguments{
\item{name}{Python variable name}
}
\description{
The data retrieved from Python has to be a list of numbers.
}

