% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_spacy}
\alias{parse_spacy}
\title{parse text using spaCy}
\usage{
parse_spacy(x, ...)
}
\arguments{
\item{x}{input text}

\item{...}{arguments passed to specific methods}
}
\value{
result marker object
}
\description{
Parse text using spaCy. The results of parsing is stored as a python object. To obtain the results in R, use functions such as get_tokens(), get_tags() etc.
\url{http://spacy.io}.
}
\examples{
\donttest{initialize_spacy_rpython()
# the result has to be "tag() is ready to run" to run the following
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
results <- parse_spacy(txt)

}
}

