% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_tag.R
\name{pos_tag}
\alias{pos_tag}
\alias{pos_tag.character}
\title{tag parts of speech using spaCy}
\usage{
pos_tag(x, tagset)

\method{pos_tag}{character}(x, tagset = c("google", "penn"))
}
\arguments{
\item{x}{input text}

\item{tagset}{character label for the tagset to use, either \code{"google"}
or \code{"penn"} to use the simplified Google tagset, or the more detailed
scheme from the Penn Treebank}
}
\value{
tagged object
}
\description{
Tokenize a text using spaCy and tag the tokens with part-of-speech tags. 
Options exist for using either the Google or Penn tagsets. See
\url{http://spacy.io}.
}
\section{Setting the path to the python executable}{
 Note that on some
  systems, notably OS X / macOS, you may have installed a different version
  of Python from that included in the base system.  OS X / macOS installs a
  slightly older version of 2.7.x by default, for instance, in
  \code{/usr/bin/python}.  Using homebrew, you may have installed a different
  version that gets placed in \code{/usr/local/bin}.  Even when this is
  working at a command line (e.g. bash in the Terminal), when called from R
  it may still look for \code{usr/bin/python}. The solution is to set the
  system variable \code{PYTHON_PATH} as in the examples below using
  \code{\link{options}}.
}
\examples{
# for my system
options(PYTHON_PATH = "/usr/local/bin")
pos_tag("This is a very short sample sentence.", tagset = "penn")
pos_tag(data_paragraph, tagset = "google")

# python inst/python/posTag.py -w < inst/extdata/test.txt 
}

