% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{print.tokenizedTexts_tagged}
\alias{print.tokenizedTexts_tagged}
\title{print a tokenizedTexts objects}
\usage{
\method{print}{tokenizedTexts_tagged}(x, sep = "_", ...)
}
\arguments{
\item{x}{a tokenizedText_tagged object created by \link{tag}}

\item{sep}{separator for printing tokens and tags, default is \code{"_"}.  If
\code{NULL}, print tokens and tags separately.}

\item{...}{further arguments passed to base print method}
}
\description{
print method for a \link{tokenize}dText object
}
\keyword{internal}

