% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_parse.R
\name{process_document}
\alias{process_document}
\title{Tokenize text using spaCy}
\usage{
process_document(x, multithread, ...)
}
\arguments{
\item{x}{input text
functionalities including the tagging, named entity recognition, dependency 
analysis. 
This slows down \code{spacy_parse()} but speeds up the later parsing. 
If FALSE, tagging, entity recognition, and dependency analysis when 
relevant functions are called.}

\item{multithread}{logical;}

\item{...}{arguments passed to specific methods}
}
\value{
result marker object
}
\description{
Tokenize text using spaCy. The results of tokenization is stored as a Python
object. To obtain the tokens results in R, use \code{get_tokens()}.
\url{http://spacy.io}.
}
\examples{
\donttest{
spacy_initialize()
# the result has to be "tag() is ready to run" to run the following
txt <- c(text1 = "This is the first sentence.\nHere is the second sentence.", 
         text2 = "This is the second document.")
results <- spacy_parse(txt)
}
}
\keyword{internal}
