% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{pydict}
\alias{pydict}
\title{Copy variables to a Python dictionary in __main__}
\usage{
pydict(x, key, dictname)
}
\arguments{
\item{x}{R object to copy}

\item{key}{Key of the object as string}

\item{name}{of the dictionary}
}
\description{
Copy variables to a Python dictionary in __main__
}

