% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pyrun}
\alias{pyrun}
\title{Run python code}
\usage{
pyrun(command)
}
\arguments{
\item{command}{Python code to execute as string}
}
\description{
Runs Python code in namespace __main__ .
}
\examples{
pyrun("print(range(5))")
}

