% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{pytype}
\alias{pytype}
\title{Get the type of Python variable}
\usage{
pytype(var)
}
\arguments{
\item{var}{name of the variable to lookup}
}
\description{
var name of the Python variable
}
\examples{
pyvar("x", 1:10)
pytype(x)
pytype(x[1])
}

