% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{pyvar}
\alias{pyvar}
\title{Copy variable from R to Python}
\usage{
pyvar(name, x)
}
\arguments{
\item{name}{variable name in python}

\item{x}{R data, numeric and character vectors and numeric matrices are currently supported.}
}
\description{
Copy variable from R to Python
}
\examples{
pyvar("x", 1:10)
pyprint(x)
pyvar("s", c("Hello", "R!")) 
pyprint(s)
df <- data.frame(x = rnorm(10), y = 1:10)
pyvar(df)
pyprint(df)
pyvar("l", list(a=3, b=1:10, z="a"))
pyprint(l)
}

