% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-methods.R
\name{quanteda-methods}
\alias{quanteda-methods}
\title{extensions of methods defind in the quanteda package}
\arguments{
\item{x}{the \code{spacyr_parsed} object returned by \code{\link{spacy_parse}}}

\item{...}{unused}
}
\description{
Extensions to quanteda functions.  You must have attached \pkg{quanteda} for these
to work.
}
\section{Usage}{

\code{docnames(x)} returns the document names

\code{ndoc(x)} returns the number of documents

\code{ntoken(x, ...)} returns the number of tokens by document

\code{ntype(x, ...)} returns the number of types (unique tokens) by document
}

\examples{
\dontrun{
require(quanteda)
spacy_initialize()
txt <- c(doc1 = "And now, now, now for something completely different.",
         doc2 = "Jack and Jill ran up the hill.")
parsed <- spacy_parse(txt)
ntype(parsed)
ntoken(parsed)
ndoc(parsed)
docnames(parsed)
}
}
