% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_condaenv_setup.R
\name{spacy_condaenv_add_models}
\alias{spacy_condaenv_add_models}
\title{Add language models to spaCy conda envirnonment}
\usage{
spacy_condaenv_add_models(models = c("en"))
}
\arguments{
\item{models}{Language model(s) for spacy to install. Example: \code{en} (English) and
\code{de} (German). Default is \code{en}. 
Possible to install multiple models at once (e.g. \code{models = c('en', 'de')}).}
}
\description{
add language models
}
\author{
Akitaka Matsuo
}
