% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_condaenv_setup.R
\name{spacy_condaenv_remove}
\alias{spacy_condaenv_remove}
\title{remove spaCy conda envirnonment}
\usage{
spacy_condaenv_remove(modify_rprofile = TRUE)
}
\arguments{
\item{modify_rprofile}{logical remove a line in .Rprofile describing the condaenv}
}
\description{
remove a conda environment for spacyr
}
\author{
Akitaka Matsuo
}
