% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_condaenv_setup.R
\name{spacy_condaenv_setup}
\alias{spacy_condaenv_setup}
\title{spaCy conda envirnonment setup}
\usage{
spacy_condaenv_setup(models = c("en"), refresh_condaenv = FALSE)
}
\arguments{
\item{models}{Language model(s) for spacy to install. Example: \code{en} (English) and
\code{de} (German). Default is \code{en}. 
Possible to install multiple models at once (e.g. \code{models = c('en', 'de')}).}

\item{refresh_condaenv}{logical. If \code{TRUE} the previous environment (if exists) will be deleted.
Default is \code{FALSE}.}
}
\description{
create a conda environment for spacyr
}
\author{
Akitaka Matsuo
}
