% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_lang_model_functions.R
\name{spacy_download_lang_model_conda}
\alias{spacy_download_lang_model_conda}
\title{install language moden in conda virtual environment}
\usage{
spacy_download_lang_model_conda(envname = "spacy_condaenv", model = "en",
  conda = "auto")
}
\arguments{
\item{envname}{Name of conda environment. Default "spacy_condaenv" (the name used by spacyr)}

\item{model}{name of the laguage model to be installed}

\item{conda}{}
}
\description{
install language moden in conda virtual environment
}
