% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_lang_model_functions.R
\name{spacy_download_lang_model_virtualenv}
\alias{spacy_download_lang_model_virtualenv}
\title{install language moden in python virtual environment}
\usage{
spacy_download_lang_model_virtualenv(envname = "spacy_virtualenv",
  model = "en", virtualenv_root = NULL)
}
\arguments{
\item{envname}{Name of virtual environment. Default "spacy_virtualenv" (the name used by spacyr)}

\item{model}{name of the laguage model to be installed}

\item{virtualenv_root}{path to the virtualenv environment to install spacy language model
if \code{NULL}, spacyr will use the default path ("~/.virtualenvs").}
}
\description{
install language moden in python virtual environment
}
