% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_langmodel_functions.R
\name{spacy_download_langmodel_virtualenv}
\alias{spacy_download_langmodel_virtualenv}
\title{Install a language model in a conda or virtual environment}
\usage{
spacy_download_langmodel_virtualenv(...)
}
\arguments{
\item{...}{not used}
}
\description{
Deprecated. \code{spacyr} now always uses a virtual environment,
making this function redundant.
}
