% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_extract_entity.R
\name{spacy_extract_entity}
\alias{spacy_extract_entity}
\title{Extract named entity}
\usage{
spacy_extract_entity(x, output = c("data.frame", "list"),
  multithread = TRUE, ...)
}
\arguments{
\item{x}{a character object or a TIF-compliant
corpus data.frame (see \url{https://github.com/ropensci/tif})}

\item{output}{type of returned object, either \code{"list"} or
\code{"data.frame"}.}

\item{multithread}{logical; If true, the processing is parallelized using pipe 
functionality of spacy (\url{https://spacy.io/api/pipe}).}

\item{...}{unused}
}
\value{
either a \code{list} or \code{data.frame} of tokens
}
\description{
This function extracts named entities from texts.
}
\examples{
\donttest{
spacy_initialize()

txt <- c(doc1 = "Natural language processing is a branch of computer science.",
         doc2 = "Paul earned a postgraduate degree from MIT.")
spacy_extract_entity(txt)
spacy_extract_entity(txt, output = "list")
}
}
