% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_extract_nounphrases.R
\name{spacy_extract_nounphrases}
\alias{spacy_extract_nounphrases}
\title{Extract noun phrases}
\usage{
spacy_extract_nounphrases(x, multithread = TRUE, value = c("list",
  "data.frame"), ...)
}
\arguments{
\item{x}{a character object, a \pkg{quanteda} corpus, or a TIF-compliant
corpus data.frame (see \url{https://github.com/ropensci/tif})}

\item{multithread}{logical; If true, the processing is parallelized using pipe 
functionality of spacy (\url{https://spacy.io/api/pipe}).}

\item{value}{type of returning object. Either \code{list} or \code{data.frame}.}

\item{...}{not used directly}
}
\value{
either \code{list} or \code{data.frame} of tokens
}
\description{
Extract noun phrases
}
\examples{
\donttest{
spacy_initialize()
txt <- "And now for something completely different."
spacy_tokenize(txt)

txt2 <- c(doc1 = "The fast cat catches mice.\\\\nThe quick brown dog jumped.", 
          doc2 = "This is the second document.",
          doc3 = "This is a \\\\\\"quoted\\\\\\" text." )
spacy_tokenize(txt2)
}
}
