% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_initialize.R
\name{spacy_finalize}
\alias{spacy_finalize}
\title{Finalize spaCy}
\usage{
spacy_finalize()
}
\description{
Finalize spaCy.
}
\details{
While running the spacy on python through R, a python process is 
always running in the backgroud and rsession will take
up a lot of memory (typically over 1.5GB). \code{spacy_finalize()} function will
finalize (i.e. terminate) the python process and free up the memory.
}
\author{
Akitaka Matsuo
}
