% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_install_model}
\alias{spacy_install_model}
\title{Install spaCy language models}
\usage{
spacy_install_model(lang_models = "en_core_web_sm")
}
\arguments{
\item{lang_models}{character; language models to be installed. Defaults
\code{en_core_web_sm} (English model). A vector of multiple model names can
be used (e.g. \code{c("en_core_web_sm", "de_core_news_sm")}). A list of
available language models and their
names is available from the \href{https://spacy.io/usage/models}{spaCy
language models} page.}
}
\value{
Invisibly returns the installation log.
}
\description{
Install spaCy language models
}
\examples{
\dontrun{
# install medium sized model
spacy_install_model("en_core_web_md")

#' # install several models with spaCy
spacy_install(lang_models = c("en_core_web_sm", "de_core_news_sm"))

# install transformer based model
spacy_install_model("en_core_web_trf")
}
}
