% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_install_virtualenv}
\alias{spacy_install_virtualenv}
\title{Install spaCy in virtualenv}
\usage{
spacy_install_virtualenv(version = "latest", lang_models = "en",
  python_version = "3.6", python_path = NULL, prompt = TRUE)
}
\arguments{
\item{version}{spaCy version to install. Specify \code{"latest"} to install
  the latest release.

  You can also provide a full major.minor.patch specification (e.g. "1.1.0")}

\item{lang_models}{Language models to be installed. Default \code{en}
(English model). A vector of multiple model names can be used (e.g.
\code{c("en", "de")})}

\item{python_version}{determine Python version for condaenv installation. 3.5
and 3.6 are available}

\item{python_path}{character; path to Python in virtualenv installation}

\item{prompt}{logical; ask whether proceed during the installation}
}
\description{
Install spaCy in virtualenv
}
