% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{spacy_parse.character}
\alias{spacy_parse.character}
\title{get tokens from a text vector}
\usage{
\method{spacy_parse}{character}(x, pos_tag = TRUE, named_entity = FALSE,
  dependency = FALSE, hash_tokens = FALSE, full_parse = FALSE, ...)
}
\arguments{
\item{x}{a text vector}

\item{pos_tag}{Logical. Part of speech tagging}

\item{named_entity}{Logical. Conduct nemed entity recognition}

\item{dependency}{Logical. Conduct dependency analysis.}

\item{hash_tokens}{Logical. Hash tokens.}
}
\value{
an tokenized text object
}
\description{
get tokens from a text vector
}
\examples{
\donttest{
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
results <- spacy_parse(txt)
tokens <- tokens(results)
}
}

