% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_uninstall}
\alias{spacy_uninstall}
\title{Unnstall spaCy conda environment}
\usage{
spacy_uninstall(conda = "auto", prompt = TRUE)
}
\arguments{
\item{conda}{Path to conda executable. Default "auto" which automatically
find the path}

\item{prompt}{logical; ask whether proceed during the installation}
}
\description{
Removes the conda environemnt craeted by spacy_install()
}
