% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_wordvector_apply.R
\name{spacy_wordvectors_apply}
\alias{spacy_wordvectors_apply}
\title{Word vector apply}
\usage{
spacy_wordvectors_apply(x, wordvectors = NULL, ...)
}
\arguments{
\item{x}{a data.frame generated from spacy_parse()}

\item{wordvectors}{returned object from spacy_wordvectors_lookup(). If \code{NULL}, spacyr trys to get vectors 
by calling \code{spacy_wordvectors_lookup()}.}
}
\value{
an unnamed matrix where each row corresponds with the word vector for input data.frame
}
\description{
To be written
}
\examples{
\donttest{
spacy_initialize(model = "en_core_web_md")

txt2 <- c(doc1 = "The fast cat catches mice.\\\\nThe quick brown dog jumped.", 
          doc2 = "This is the second document.",
          doc3 = "This is a \\\\\\"quoted\\\\\\" text." )
spacy_output <- spacy_parse(txt2)
wordvector_matrix <- spacy_wordvectors_apply(spacy_output)
}
}
