% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_wordvector_lookup.R
\name{spacy_wordvectors_lookup}
\alias{spacy_wordvectors_lookup}
\title{Vector look up}
\usage{
spacy_wordvectors_lookup(x, ...)
}
\arguments{
\item{x}{a data.frame generated from spacy_parse() or from spacy_tokenize() or 
a vocabulary vector (a character vector with unique elements)}
}
\value{
a list of word vectors, named by tokens
}
\description{
To be written
}
\examples{
\donttest{
spacy_initialize()
txt <- "And now for something completely different."
spacy_tokenize(txt)

txt2 <- c(doc1 = "The fast cat catches mice.\\\\nThe quick brown dog jumped.", 
          doc2 = "This is the second document.",
          doc3 = "This is a \\\\\\"quoted\\\\\\" text." )
spacy_tokenize(txt2)
}
}
