% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/package.R
\docType{package}
\name{spacyr-package}
\alias{spacyr}
\alias{spacyr-package}
\alias{spacyr}
\alias{spacyr-package}
\title{An R wrapper to the spaCy NLP system}
\description{
An R wrapper to the Python (Cython) spaCy NLP system, from http://spacy.io. 
Nicely integrated with \pkg{quanteda}.

Spacy for R
}
\details{
To use \pkg{spacyr}, you must first have Python and spaCy installed on your
system.  See the installation instructions from the links below.
}
\references{
\url{http://spacy.io}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/kbenoit/spacyr}
  \item Report bugs at \url{https://github.com/kbenoit/spacyr/issues}
}

}
\author{
Ken Benoit and Akitaka Matsuo
}
