% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{summary.tokenizedTexts_tagged}
\alias{summary.tokenizedTexts_tagged}
\title{summarize a tagged tokenizedTexts object}
\usage{
\method{summary}{tokenizedTexts_tagged}(object, ...)
}
\arguments{
\item{object}{tokenizedTexts_tagged object to be summarized}

\item{...}{unused}
}
\description{
Generate frequency counts of POS by document, returning a data.frame.
}

