% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\alias{tag.character}
\alias{tag.corpus}
\title{tag parts of speech using spaCy}
\usage{
tag(x, ...)

\method{tag}{corpus}(x, ...)

\method{tag}{character}(x, tagset = c("google", "penn"), ...)
}
\arguments{
\item{x}{input text}

\item{...}{arguments passed to specific methods}

\item{tagset}{character label for the tagset to use, either \code{"google"} 
or \code{"penn"} to use the simplified Google tagset, or the more detailed 
scheme from the Penn Treebank.}
}
\value{
tagged object
}
\description{
Tokenize a text using spaCy and tag the tokens with part-of-speech tags. 
Options exist for using either the Google or Penn tagsets. See 
\url{http://spacy.io}.
}
\section{Setting the path to the python executable}{
 Note that on some 
  systems, notably OS X / macOS, you may have installed a different version 
  of Python from that included in the base system.  OS X / macOS installs a 
  slightly older version of 2.7.x by default, for instance, in 
  \code{/usr/bin/python}.  Using homebrew, you may have installed a different
  version that gets placed in \code{/usr/local/bin}.  Even when this is 
  working at a command line (e.g. bash in the Terminal), when called from R 
  it may still look for \code{usr/bin/python}. The solution is to set the 
  system variable \code{PYTHON_PATH} as in the examples below using 
  \code{\link{options}}.
}
\examples{
# for my system
options(PYTHON_PATH = "/usr/local/bin")
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
tag(txt)
tag(txt, tagset = "penn")

# more extensive texts
tag(data_paragraph)
tag(data_sentences[1:10])
}

