% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tag_new}
\alias{tag_new}
\title{tag parts of speech using spaCy via rPython}
\usage{
tag_new(spacy_out, tagset = c("penn", "google"))
}
\arguments{
\item{spacy_out}{a spacy_out object generated from \code{parse_spacy}}

\item{tagset}{character label for the tagset to use, either \code{"google"} 
or \code{"penn"} to use the simplified Google tagset, or the more detailed 
scheme from the Penn Treebank.}
}
\value{
a tokenizedTexts_tagged object
}
\description{
Tokenize a text using spaCy and tag the tokens with part-of-speech tags. 
Options exist for using either the Google or Penn tagsets. See 
\url{http://spacy.io}.
}

