% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag_rPython}
\alias{tag_rPython}
\title{tag parts of speech using spaCy via rPython}
\usage{
tag_rPython(x, tagset = c("google", "penn"), ...)
}
\arguments{
\item{x}{input text}

\item{tagset}{character label for the tagset to use, either \code{"google"} 
or \code{"penn"} to use the simplified Google tagset, or the more detailed 
scheme from the Penn Treebank.}

\item{...}{arguments passed to specific methods}
}
\value{
tagged object
}
\description{
Tokenize a text using spaCy and tag the tokens with part-of-speech tags. 
Options exist for using either the Google or Penn tagsets. See 
\url{http://spacy.io}.
}
\examples{
\donttest{initialize_spacy_rpython()
# the result has to be "tag() is ready to run" to run the following
txt <- c(text1 = "This is the first sentence.\\n Here is the second sentence.", 
         text2 = "This is the second document.")
tag_rPython(txt)
tag_rPython(txt, tagset = "penn")

# more extensive texts
tag_rPython(data_paragraph)
tag_rPython(data_sentences[1:4])
}
}

