% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tokens}
\alias{tokens}
\alias{tokens.spacy_out}
\title{get tokens from tokenized texts}
\usage{
tokens(x, ...)

\method{tokens}{spacy_out}(spacy_out)
}
\arguments{
\item{spacy_out}{a spacy_out object}
}
\value{
an tokenized text object in tokenizedText_spacyr class
}
\description{
get tokens from tokenized texts
}
\section{Methods (by class)}{
\itemize{
\item \code{spacy_out}: Get tokens using spacy_out object
}}
\examples{
\donttest{
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
results <- spacy_parse(txt)
tokens <- tokens(results)
}
}

