% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tokens_named_entities}
\alias{tokens_named_entities}
\title{Title}
\usage{
tokens_named_entities(spacy_out)
}
\arguments{
\item{spacy_out}{a spacy_out object}
}
\value{
list of named entities in texts
}
\description{
Title
}

