% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tokens_out}
\alias{tokens_out}
\title{create quanteda tokenized text object from spacy results}
\usage{
tokens_out(dt)
}
\arguments{
\item{dt}{a data.table output from spacy_parse()}
}
\value{
a tokenizedText object
}
\description{
create quanteda tokenized text object from spacy results
}

