% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{topy.numeric}
\alias{topy.numeric}
\title{Push data to python __main__ namespace}
\usage{
\method{topy}{numeric}(x, name)
}
\arguments{
\item{x}{Numeric vector to copy to Python}

\item{name}{Python variable name as string}
}
\description{
Push data to python __main__ namespace
}

