% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{all_named_entities}
\alias{all_named_entities}
\title{Obtain a data of all named entities in spacy out data.table}
\usage{
all_named_entities(dt)
}
\arguments{
\item{dt}{a data table object from}
}
\value{
a data.table of all named entities
}
\description{
Extended description to go here.
}

