% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-functions.R
\name{entity_consolidate}
\alias{entity_consolidate}
\title{Consolidate named entities in parsed output}
\usage{
entity_consolidate(spacy_result, concatenator = "_")
}
\arguments{
\item{spacy_result}{a \code{data.frame} from \code{\link{spacy_parse}}.}

\item{concatenator}{the character used to concatenator elements of multi-word named entities}
}
\value{
A modified \code{data.frame} of spacy outputs, where the named entities have been 
combined into a single "token".
}
\description{
\code{entity_consolidate} concatinate named entity in a table of
results of \code{spacy_parse}
}
\examples{
\donttest{
spacy_initialize()

parsed_sentences <- spacy_parse(data_char_sentences, entity = TRUE)
parsed_sentences_consolidated <- entity_consolidate(parsed_sentences)
head(parsed_sentences_consolidated, 30)
}
}
