% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_initialize.R
\name{find_spacy}
\alias{find_spacy}
\title{Find spaCy}
\usage{
find_spacy(model = "en_core_web_sm", ask)
}
\arguments{
\item{model}{name of the language model}

\item{ask}{logical; if \code{FALSE}, use the first spaCy installation found;
if \code{TRUE}, list available spaCy installations and prompt the user
for which to use. If another (e.g. \code{python_executable}) is set, then
this value will always be treated as \code{FALSE}.}
}
\value{
spacy_python
}
\description{
Locate the user's version of Python for which spaCy installed.
}
\keyword{internal}
