% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_attrs}
\alias{get_attrs}
\title{get arbitrary attributes from spacy output}
\usage{
get_attrs(spacy_out, attr_name)
}
\arguments{
\item{spacy_out}{a spacy_out object}

\item{attr_name}{name of spacy token attributes to extract}
}
\value{
a list of attributes
}
\description{
get arbitrary attributes from spacy output
}

