% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-extractor-functions.R
\name{get_dependency}
\alias{get_dependency}
\title{Title}
\usage{
get_dependency(spacy_out)
}
\arguments{
\item{spacy_out}{a spacy_out object}
}
\value{
data.frame of dependency relations
}
\description{
Title
}
\keyword{internal}

