% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_tags}
\alias{get_tags}
\title{get tags from spaCy output}
\usage{
get_tags(spacy_out, tagset = c("google", "penn"))
}
\arguments{
\item{spacy_out}{a spacy_out object}

\item{tagset}{character label for the tagset to use, either \code{"google"} 
or \code{"penn"} to use the simplified Google tagset, or the more detailed 
scheme from the Penn Treebank.}
}
\value{
a list of tags
}
\description{
get tags from spaCy output
}
\examples{
\donttest{
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
results <- parse_spacy(txt)
get_tags(results)
}
}

