% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spacy.R
\name{initialize_spacy_rpython}
\alias{initialize_spacy_rpython}
\title{Initialize spaCy via rPython}
\usage{
initialize_spacy_rpython()
}
\description{
Initialize spaCy using an alternative method via the rPython package.
}
\author{
Akitaka Matsuo
}

