% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_tag.R
\name{pos_tag}
\alias{pos_tag}
\alias{pos_tag.character}
\title{tag parts of speech using spaCy}
\usage{
pos_tag(x, tagset)

\method{pos_tag}{character}(x, tagset = c("google", "penn"))
}
\arguments{
\item{x}{input text}

\item{tagset}{character label for the tagset to use, either \code{"google"} or \code{"penn"} to use
the simplified Google tagset, or the more detailed scheme from the Penn Treebank}
}
\value{
tagged object
}
\description{
Tokenize a text using spaCy and tag the tokens with part-of-speech tags. 
Options exist for using either the Google or Penn tagsets.
See \url{http://spacy.io}.
}
\examples{
# fails on my system:
pos_tag("This is a very short sample sentence.", tagset = "google")

# Note: This works from my command line:
# python inst/python/posTag.py -w < inst/extdata/test.txt 
}

