% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{pyprint}
\alias{pyprint}
\title{Print python object}
\usage{
pyprint(x)
}
\arguments{
\item{name}{name of the python object}
}
\description{
Print python object
}
\examples{
pyvar("x", 1:10)
pyprint(x)
pyprint("dir()") #You can quote Python commands
}

