% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_lang_model_functions.R
\name{spacy_download_lang_model_conda}
\alias{spacy_download_lang_model_conda}
\alias{spacy_download_lang_model_virtualenv}
\title{Install a language model in a virtual environment}
\usage{
spacy_download_lang_model_conda(envname = "spacy_condaenv", model = "en",
  conda = "auto")

spacy_download_lang_model_virtualenv(envname = "spacy_virtualenv",
  model = "en", virtualenv_root = NULL)
}
\arguments{
\item{envname}{name of the virtual environment}

\item{model}{name of the language model to be installed}

\item{conda}{Path to conda executable.  Default \code{"auto"} which
automatically finds the path.}

\item{virtualenv_root}{path to the virtualenv environment to install spaCy
language model. If \code{NULL}, the default path \code{"~/.virtualenvs"}
will be used.}
}
\description{
Installs one or more language models in a conda or virtualenv Python virtual
environment as installed by \code{\link{spacy_install}}.
}
