% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_initialize.R
\name{spacy_finalize}
\alias{spacy_finalize}
\title{Finalize spaCy}
\usage{
spacy_finalize()
}
\description{
While running the spacy on python through R, a python process is 
always running in the backgroud and rsession will take
up a lot of memory (typically over 1.5GB). \code{spacy_finalize()} terminates the 
Python process and frees up the memory it was using.
}
\author{
Akitaka Matsuo
}
