% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_initialize.R
\name{spacy_initialize}
\alias{spacy_initialize}
\title{Initialize spaCy}
\usage{
spacy_initialize(lang = "en")
}
\arguments{
\item{lang}{Language package for loading spacy. Either \code{en} (English) or 
\code{de} (German). Default is \code{en}.}
}
\description{
Initialize spaCy to call from R.
}
\author{
Akitaka Matsuo
}
