% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_initialize.R
\name{spacy_initialize}
\alias{spacy_initialize}
\title{Initialize spaCy via rPython}
\usage{
spacy_initialize()
}
\description{
Initialize spaCy using an alternative method via the rPython package.
}
\author{
Akitaka Matsuo
}

