% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_initialize.R
\name{spacy_initialize}
\alias{spacy_initialize}
\title{Initialize spaCy}
\usage{
spacy_initialize(model = "en_core_web_sm", entity = TRUE, ...)
}
\arguments{
\item{model}{Language package for loading spaCy. Example: \code{en_core_web_sm} (English) and
\code{de_core_web_sm} (German). Default is \code{en_core_web_sm}.}

\item{entity}{logical; if \code{FALSE} is selected, named entity recognition
is turned off in spaCy. This will speed up the parsing as it will exclude
\code{ner} from the pipeline. For details of spaCy pipeline, see
\url{https://spacy.io/usage/processing-pipelines}. The option \code{FALSE}
is available only for spaCy version 2.0.0 or higher.}

\item{...}{not used.}
}
\description{
Initialize spaCy to call from R.
}
\author{
Akitaka Matsuo, Johannes B. Gruber
}
