% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_initialize.R
\name{spacy_initialize}
\alias{spacy_initialize}
\title{Initialize spaCy}
\usage{
spacy_initialize(model = "en", python_executable = NULL, ask = FALSE,
  virtualenv = NULL, condaenv = NULL)
}
\arguments{
\item{model}{Language package for loading spacy. Example: \code{en} (English) and
\code{de} (German). Default is \code{en}.}

\item{python_executable}{the full path to the python excutable, for which spaCy is installed}

\item{ask}{logical; if \code{FALSE}, use the first spaCy installation found; 
if \code{TRUE}, list available spaCy installations and prompt the user 
for which to use. If another (e.g. \code{python_executable}) is set, then 
this value will always be treated as \code{FALSE}.}

\item{virtualenv}{set a path to the python virtual environment with spaCy installed
Example: \code{virtualenv = "~/myenv"}}

\item{condaenv}{set a path to the anaconda virtual environment with spaCy installed
Example: \code{condalenv = "myenv"}}
}
\description{
Initialize spaCy to call from R.
}
\author{
Akitaka Matsuo
}
