% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_install}
\alias{spacy_install}
\title{Install spacy}
\usage{
spacy_install(method = c("virtualenv", "conda", "system"), conda = "auto",
  version = "latest", lang_models = "en", python_version = "3.6",
  python_path = NULL, restart_session = TRUE)
}
\arguments{
\item{method}{}

\item{conda}{the version of conda to be used}

\item{version}{Spacy version to install. Specify "latest" to install
  the latest release. 

  You can also provide a full major.minor.patch specification (e.g. "1.1.0")}

\item{lang_models}{Language models to be installed. Default \code{en} (English model). 
A vector of multiple model names can be used (e.g. \code{c('en', 'de')})}

\item{python_version}{determine python version for condaenv installation. 3.5 and 3.6 are available}

\item{python_path}{supply path to python in virtualenv installation}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}
}
\description{
Install spacy
}
