% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_install}
\alias{spacy_install}
\title{Install spaCy in conda or virtualenv environment}
\usage{
spacy_install(
  version = "latest",
  lang_models = "en_core_web_sm",
  ask = interactive(),
  force = FALSE,
  ...
)
}
\arguments{
\item{version}{character; spaCy version to install (see details).}

\item{lang_models}{character; language models to be installed. Defaults
\code{en_core_web_sm} (English model). A vector of multiple model names can
be used (e.g. \code{c("en_core_web_sm", "de_core_news_sm")}). A list of
available language models and their
names is available from the \href{https://spacy.io/usage/models}{spaCy
language models} page.}

\item{ask}{logical; ask whether to proceed during the installation. By
default, questions are only asked in interactive sessions.}
}
\description{
Install spaCy in a self-contained environment, including
  specified language models.
}
\examples{
\dontrun{
# install the latest version of spaCy
spacy_install()

# update spaCy
spacy_install(force = TRUE)

# install an older version
spacy_install("3.1.0")

# install with GPU enabled
spacy_install("cuda-autodetect")

# install on Apple ARM processors
spacy_install("apple")

# install several models with spaCy
spacy_install(lang_models = c("en_core_web_sm", "de_core_news_sm"))

# install spaCy to an existing virtual environment
Sys.setenv(RETICULATE_PYTHON = "path/to/python")
spacy_install()
}

}
\seealso{
\code{\link{spacy_install_model}}
}
