% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_install_virtualenv}
\alias{spacy_install_virtualenv}
\title{Install spaCy to a virtual environment}
\usage{
spacy_install_virtualenv(...)
}
\arguments{
\item{...}{not used}
}
\description{
Deprecated. \code{spacy_install} now installs to a virtual environment by default.
}
