% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{spacy_parse.character}
\alias{spacy_parse.character}
\title{get tokens from a text vector}
\usage{
\method{spacy_parse}{character}(x, pos_tag = TRUE, named_entity = FALSE,
  dependency = FALSE, hash_tokens = FALSE, full_parse = FALSE, ...)
}
\arguments{
\item{x}{a character object}

\item{pos_tag}{logical; if \code{TRUE}, tag parts of speech}

\item{named_entity}{logical; if \code{TRUE}, report named entities}

\item{dependency}{logical; if \code{TRUE}, analyze and return dependencies}

\item{hash_tokens}{logical; if \code{TRUE}, hash the tokens}

\item{full_parse}{logical; if \code{TRUE}, conduct the one-shot parse 
regardless of the value of other parameters. This  option exists because
the parsing outcomes of named entities are slightly different different
when named entities are run separately from the parsing.}

\item{...}{additional arguments required for methods passing}
}
\value{
an tokenized text object
}
\description{
Full description awaits.
}
\examples{
\donttest{
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
results <- spacy_parse(txt)
tokens <- tokens(results)
}
}

