% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{spacy_parse}
\alias{spacy_parse}
\alias{spacy_parse.corpus}
\title{Tokens}
\usage{
spacy_parse(x, ...)

\method{spacy_parse}{corpus}(x, ...)
}
\arguments{
\item{x}{objet to be parsed}

\item{...}{additional arguments not used}
}
\description{
Tokens
}

