% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_uninstall}
\alias{spacy_uninstall}
\title{Unnstall spaCy conda environment}
\usage{
spacy_uninstall(conda = "auto", envname = "spacy_condaenv")
}
\arguments{
\item{conda}{Path to conda executable. Default "auto" which automatically
find the path}

\item{envname}{character; name of conda environent to remove}
}
\description{
Removes the conda environemnt craeted by spacy_install()
}
