% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_uninstall}
\alias{spacy_uninstall}
\title{Uninstall spaCy conda environment}
\usage{
spacy_uninstall(conda = "auto", prompt = TRUE,
  envname = "spacy_condaenv")
}
\arguments{
\item{conda}{Path to conda executable, default to "auto" which automatically
finds the path}

\item{prompt}{logical; ask whether to proceed during the installation}

\item{envname}{character; name of conda environment to remove}
}
\description{
Removes the conda environment created by spacy_install()
}
