% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_uninstall}
\alias{spacy_uninstall}
\title{Uninstall the spaCy environment}
\usage{
spacy_uninstall(confirm = interactive())
}
\arguments{
\item{confirm}{logical; confirm before uninstalling spaCy?}
}
\description{
Removes the virtual environment created by spacy_install()
}
