% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_install.R
\name{spacy_upgrade}
\alias{spacy_upgrade}
\title{Install spaCy in conda environment}
\usage{
spacy_upgrade(conda = "auto", envname = "spacy_condaenv",
  lang_models = "en")
}
\arguments{
\item{conda}{Path to conda executable. Default "auto" which automatically
find the path}

\item{envname}{character; name of conda environment to upgrate spaCy}

\item{lang_models}{Language models to be upgraded. Default NULL (No upgrade). 
A vector of multiple model names can be used (e.g. \code{c("en", "de")})}
}
\description{
Install spaCy in conda environment
}
