% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacyr-package.R
\docType{package}
\name{spacyr-package}
\alias{spacyr}
\alias{spacyr-package}
\title{An R wrapper to the spaCy NLP system}
\description{
An R wrapper to the spaCy NLP system
}
\author{
Ken Benoit and Akitaka Matsuo

An R wrapper to the Python (Cython) spaCy NLP system, 
from http://spacy.io.  Nicely integrated with \pkg{quanteda}.
}

