% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{spacyr-package}
\alias{spacyr}
\alias{spacyr-package}
\title{An R wrapper to the spaCy NLP system}
\description{
An R wrapper to the Python (Cython) spaCy NLP system, from
\url{http://spacy.io}. Nicely integrated with \pkg{quanteda}.  \pkg{spacyr}
is designed to provide easy access to the powerful functionality of spaCy, in
a simple format.
}
\details{
To use \pkg{spacyr}, you must first have Python and spaCy installed on your 
system.  See the installation instructions from 
\url{http://github.com/kbenoit/spacyr}.
}
\references{
\url{http://spacy.io}, \url{http://github.com/kbenoit/spacyr}.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/quanteda/spacyr}
  \item Report bugs at \url{https://github.com/quanteda/spacyr/issues}
}

}
\author{
Ken Benoit and Akitaka Matsuo
}
