% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\alias{tag.character}
\alias{tag.corpus}
\title{tag parts of speech using spaCy}
\usage{
tag(x, ...)

\method{tag}{corpus}(x, ...)

\method{tag}{character}(x, tagset = c("google", "penn"), ...)
}
\arguments{
\item{x}{input text}

\item{...}{arguments passed to specific methods}

\item{tagset}{character label for the tagset to use, either \code{"google"} 
or \code{"penn"} to use the simplified Google tagset, or the more detailed 
scheme from the Penn Treebank.}
}
\value{
tagged object
}
\description{
Tokenize a text using spaCy and tag the tokens with part-of-speech tags. 
Options exist for using either the Google or Penn tagsets. See 
\url{http://spacy.io}.
}
\examples{
\donttest{initialize_spacy()
# the result has to be "tag() is ready to run" to run the following
txt <- c(text1 = "This is the first sentence.\\nHere is the second sentence.", 
         text2 = "This is the second document.")
tag(txt)
tag(txt, tagset = "penn")

# more extensive texts
tag(data_paragraph)
tag(data_sentences[1:10])
}
}

