% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-UNUSED.R
\name{tokens_tags_out}
\alias{tokens_tags_out}
\title{create quanteda tokenized text object from spacy results with tag}
\usage{
tokens_tags_out(dt, tagset = "penn")
}
\arguments{
\item{dt}{a data.table output from \code{\link{spacy_parse}}}

\item{tagset}{character label for the tagset to use, either \code{"google"} 
or \code{"penn"} to use the simplified Google tagset, or the more detailed 
scheme from the Penn Treebank.}
}
\value{
a tokenizedText object
}
\description{
Full description goes here.
}
\keyword{internal}

